pyproject.toml
[project]
name = "familybot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.1.1",
]
